Let's add a sales chart program!
Now we have a sales chart repository!
Ok, so this does exactly what the board is there for, except with the
extra ability to update every 15 seconds to whatever's in our sales data
table. We have the ability to add dates for weird days, we can display
seperated data for ebay, amazon, our website and direct sales, and we
can set the view to full screen for a public display without visible
settings.